requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9

pymupdf==1.24.14
pillow==10.4.0

pytesseract==0.3.13

transformers==4.48.2
torch==2.5.1

sentence-transformers==3.3.1
numpy==2.1.3

pymongo==4.10.1

easyocr==1.7.0
paddleocr==3.3.0
